🗺️ Atlas: [Architectural change] Split conversion.rs blob into structured module
This commit splits the monolithic `src/semantic/conversion.rs` (~2600 lines) into a properly structured module directory. The logic is separated into `classification.rs` (statement intent diagnosis) and `extraction.rs` (value extraction). Shared utilities are placed in `shared.rs` to ensure a strict unidirectional dependency graph and eliminate circular dependencies between the two core logic files. The `mod.rs` acts as a facade, preserving the existing `convert_assembled_to_analyzed` public interface.
