+++ src/semantic/conversion/mod.rs
@@ -1,0 +1,70 @@
+//! Conversion from assembled statements to analyzed statements
+//!
+//! This module acts as the "interpreter" of the assembled semantic structure.
+//! While the [`Assembler`](crate::semantic::Assembler) ensures grammatical correctness (Subject-Verb agreement),
+//! this module assigns *meaning* to the grammatical structures.
+//!
+//! # The Interpreter Pattern
+//!
+//! The conversion process is essentially an interpretation step. It takes a
+//! grammatically valid but semantically ambiguous "Assembled Statement" and
+//! converts it into a typed, unambiguous "Analyzed Statement" (part of the HIR).
+//!
+//! This is where "word order independence" meets "semantic meaning".
+//!
+//! # Pattern Detection Strategy
+//!
+//! The [`classify_assembled_statement`] function uses a combination of strategies to
+//! understand the statement's intent, checking patterns in a specific heuristic order:
+//!
+//! 1. **Pattern Delegation**: Complex patterns are delegated first.
+//!    - **Iterator Chains**: `detect_iterator_pattern` (e.g., `list doubling print`).
+//!    - **Property Access**: `classify_property_access_print` (e.g., `user.name print`).
+//!    - **Struct Instantiation**: `try_parse_struct_instantiation` (e.g., `x new User ... let`).
+//!    - **Function Calls**: `classify_function_call` (e.g., `my_func arg1 arg2 call`).
+//!
+//! 2. **Verb-Based Classification**: If no complex pattern matches, the main verb drives the logic.
+//!    - **Binding** (`ἔστω`): `let x = value`.
+//!    - **Assignment** (`γίγνεται`): `x = value`.
+//!    - **Collection Ops** (`ὠθεῖ`, `ἕλκεται`, `τίθησι`): `push`, `pop`, `insert`.
+//!    - **Print** (`λέγε`, `γράφε`): `println!`.
+//!    - **Query** (`?`): Expressions ending in `?`.
+//!
+//! 3. **Expression Fallback**: If no verb implies a statement, it is treated as a pure expression.
+//!    - **Operations**: `1 + 2`.
+//!    - **Try/Propagate**: `expr;` (becomes `expr?`).
+
+pub(crate) mod classification;
+pub(crate) mod extraction;
+pub(crate) mod shared;
+
+pub(crate) use classification::classify_assembled_statement;
+pub(crate) use extraction::extract_value;
+
+use super::assembly::AssembledStatement;
+use super::model::AnalyzedStatement;
+use super::resolver::Scope;
+use crate::errors::GlossaError;
+
+/// Convert an AssembledStatement to an AnalyzedStatement
+///
+/// This is the main entry point for lowering the "Assembled" semantic model (slot-based)
+/// to the "Analyzed" model (HIR/AST-like).
+///
+/// Evaluates and translates a grammatically sound statement into the semantically typed AST.
+///
+/// This serves as the top-level interpreter connecting the raw output of the
+/// [`crate::semantic::Assembler`] (`AssembledStatement`) with the High-Level Intermediate
+/// Representation (`AnalyzedStatement`). It assigns concrete meaning to grammatical roles
+/// (e.g., assigning a Subject the role of "Variable Name").
+///
+/// # Arguments
+///
+/// * `asm_stmt` - The assembled statement from the `Assembler`.
+/// * `scope` - The current semantic scope (for variable lookup and definition).
+pub fn convert_assembled_to_analyzed(
+    asm_stmt: &AssembledStatement,
+    scope: &mut Scope,
+) -> Result<AnalyzedStatement, GlossaError> {
+    classify_assembled_statement(asm_stmt, scope)
+}
